improve channel utilization
